[Paths] Added Paths to manage file path creation
Reverted logging back to get argument expansion to work. Need to review that later. Scene file clean up. Added Paths namespace to manage file and directory path management for simplicity. Added placeholder ComponentLight.

+++ Framework/Source/QF_AssetsTypes.h
@@ -17,7 +17,7 @@ namespace QwerkE {
 		obj,
 	};
 
-	QC_ENUM(eAssetType, int,
+	QC_ENUM(eAssetType, u8,
 		// Invalid, Use AssetHandle::INVALID_HANDLE instead
 		Texture
 	);
